refactor(config): derive messaging types from schemas
