+++ src/config/zod-schema.core.ts
@@ -11,10 +11,17 @@ import type { ModelCompatConfig } from "./types.models.js";
 import { MODEL_APIS, MODEL_THINKING_FORMATS } from "./types.models.js";
 import { ENV_SECRET_REF_ID_RE } from "./types.secrets.js";
 import { createAllowDenyChannelRulesSchema } from "./zod-schema.allowdeny.js";
+import { DmConfigSchema } from "./zod-schema.messages.js";
 import { SecretInputSchema } from "./zod-schema.secret-input.js";
 import { sensitive } from "./zod-schema.sensitive.js";
 
 export { isBuiltInModelProviderOverlayId } from "./model-provider-config.js";
+export {
+  DmConfigSchema,
+  GroupChatSchema,
+  MentionPatternsPolicySchema,
+  ProviderCommandsSchema,
+} from "./zod-schema.messages.js";
 export { SecretInputSchema, SecretRefSchema } from "./zod-schema.secret-input.js";
 
 const WINDOWS_ABS_PATH_PATTERN = /^[A-Za-z]:[\\/]/;
@@ -542,47 +549,6 @@ export const ModelsConfigSchema = z
   .strict()
   .optional();
 
-const VisibleRepliesValueSchema = z.enum(["automatic", "message_tool"]);
-const AmbientGroupInboundSchema = z.enum(["user_request", "room_event"]);
-
-export const VisibleRepliesSchema = z
-  .union([VisibleRepliesValueSchema, z.boolean()])
-  .overwrite((value) => {
-    if (value === true) {
-      return "automatic";
-    }
-    if (value === false) {
-      return "message_tool";
-    }
-    return value;
-  });
-
-const MentionPatternsModeSchema = z.union([z.literal("allow"), z.literal("deny")]);
-
-export const MentionPatternsPolicySchema = z
-  .object({
-    mode: MentionPatternsModeSchema.optional(),
-    allowIn: z.array(z.string()).optional(),
-    denyIn: z.array(z.string()).optional(),
-  })
-  .strict();
-
-export const GroupChatSchema = z
-  .object({
-    mentionPatterns: z.array(z.string()).optional(),
-    historyLimit: z.number().int().min(0).optional(),
-    unmentionedInbound: AmbientGroupInboundSchema.optional(),
-    visibleReplies: VisibleRepliesSchema.optional(),
-  })
-  .strict()
-  .optional();
-
-export const DmConfigSchema = z
-  .object({
-    historyLimit: z.number().int().min(0).optional(),
-  })
-  .strict();
-
 export const IdentitySchema = z
   .object({
     name: z.string().optional(),
@@ -593,13 +559,6 @@ export const IdentitySchema = z
   .strict()
   .optional();
 
-const QueueModeSchema = z.union([
-  z.literal("steer"),
-  z.literal("followup"),
-  z.literal("collect"),
-  z.literal("interrupt"),
-]);
-const QueueDropSchema = z.union([z.literal("old"), z.literal("new"), z.literal("summarize")]);
 export const ReplyToModeSchema = z.union([
   z.literal("off"),
   z.literal("first"),
@@ -813,45 +772,6 @@ export const requireAllowlistAllowFrom = (params: {
 
 export const MSTeamsReplyStyleSchema = z.enum(["thread", "top-level"]);
 
-const QueueModeBySurfaceSchema = z
-  .object({
-    whatsapp: QueueModeSchema.optional(),
-    telegram: QueueModeSchema.optional(),
-    discord: QueueModeSchema.optional(),
-    irc: QueueModeSchema.optional(),
-    googlechat: QueueModeSchema.optional(),
-    slack: QueueModeSchema.optional(),
-    mattermost: QueueModeSchema.optional(),
-    signal: QueueModeSchema.optional(),
-    imessage: QueueModeSchema.optional(),
-    msteams: QueueModeSchema.optional(),
-    webchat: QueueModeSchema.optional(),
-    matrix: QueueModeSchema.optional(),
-  })
-  .strict()
-  .optional();
-
-const DebounceMsBySurfaceSchema = z.record(z.string(), z.number().int().nonnegative()).optional();
-
-export const QueueSchema = z
-  .object({
-    mode: QueueModeSchema.optional(),
-    byChannel: QueueModeBySurfaceSchema,
-    debounceMsByChannel: DebounceMsBySurfaceSchema,
-    cap: z.number().int().positive().optional(),
-    drop: QueueDropSchema.optional(),
-  })
-  .strict()
-  .optional();
-
-export const InboundDebounceSchema = z
-  .object({
-    debounceMs: z.number().int().nonnegative().optional(),
-    byChannel: DebounceMsBySurfaceSchema,
-  })
-  .strict()
-  .optional();
-
 export const HexColorSchema = z.string().regex(/^#?[0-9a-fA-F]{6}$/, "expected hex color (RRGGBB)");
 
 export const ExecutableTokenSchema = z
@@ -965,13 +885,4 @@ export const ToolsLinksSchema = z
   .strict()
   .optional();
 
-export const NativeCommandsSettingSchema = z.union([z.boolean(), z.literal("auto")]);
-
-export const ProviderCommandsSchema = z
-  .object({
-    native: NativeCommandsSettingSchema.optional(),
-    nativeSkills: NativeCommandsSettingSchema.optional(),
-  })
-  .strict()
-  .optional();
 /* oxlint-disable max-lines -- TODO: split this grandfathered oversized file. */
