+++ src/config/zod-schema.session.ts
@@ -1,42 +1,10 @@
 // Defines session-related Zod schema fragments for config parsing.
 import { z } from "zod";
 import { ElevatedAllowFromSchema } from "./zod-schema.agent-runtime.js";
-import {
-  GroupChatSchema,
-  InboundDebounceSchema,
-  NativeCommandsSettingSchema,
-  QueueSchema,
-  VisibleRepliesSchema,
-} from "./zod-schema.core.js";
+import { NativeCommandsSettingSchema } from "./zod-schema.messages.js";
 
+export { MessagesSchema } from "./zod-schema.messages.js";
 export { SessionSchema } from "./zod-schema.session-config.js";
-
-const ResponseUsageModeSchema = z.enum(["on", "off", "tokens", "full"]);
-
-export const MessagesSchema = z
-  .object({
-    visibleReplies: VisibleRepliesSchema.optional(),
-    responsePrefix: z.string().optional(),
-    usageTemplate: z.union([z.string(), z.record(z.string(), z.unknown())]).optional(),
-    responseUsage: z
-      .union([ResponseUsageModeSchema, z.record(z.string(), ResponseUsageModeSchema)])
-      .optional(),
-    groupChat: GroupChatSchema,
-    queue: QueueSchema,
-    inbound: InboundDebounceSchema,
-    ackReaction: z.string().optional(),
-    ackReactionScope: z
-      .enum(["group-mentions", "group-all", "direct", "all", "off", "none"])
-      .optional(),
-    statusReactions: z
-      .object({
-        enabled: z.boolean().optional(),
-      })
-      .strict()
-      .optional(),
-  })
-  .strict()
-  .optional();
 
 export const CommandsSchema = z
   .object({
@@ -55,11 +23,4 @@ export const CommandsSchema = z
   })
   .strict()
   .optional()
-  .default(
-    () =>
-      ({
-        native: "auto",
-        nativeSkills: "auto",
-        restart: true,
-      }) as const,
-  );
+  .default(() => ({ native: "auto", nativeSkills: "auto", restart: true }) as const);
